refactor: consolidate
