+++ packages/react-native-screen-transitions/src/shared/hooks/lifecycle/use-close-transition.ts
@@ -10,7 +10,7 @@ import { useStackCoreContext } from "../../providers/stack/core.provider";
 import { useManagedStackContext } from "../../providers/stack/managed.provider";
 import type { AnimationStoreMap } from "../../stores/animation.store";
 import { StackType } from "../../types/stack.types";
-import { startScreenTransition } from "../../utils/animation/start-screen-transition";
+import { animateToProgress } from "../../utils/animation/animate-to-progress";
 import { resetStoresForScreen } from "../../utils/reset-stores-for-screen";
 import { useSharedValueState } from "../reanimated/use-shared-value-state";
 import useStableCallback from "../use-stable-callback";
@@ -49,7 +49,7 @@ const useManagedClose = ({
 			if (!keys?.includes(current.route.key)) return;
 
 			runOnJS(activate)();
-			startScreenTransition({
+			animateToProgress({
 				target: "close",
 				spec: current.options.transitionSpec,
 				animations,
@@ -96,7 +96,7 @@ const useNativeStackClose = ({
 		e.preventDefault();
 		activate();
 
-		startScreenTransition({
+		animateToProgress({
 			target: "close",
 			spec: current.options.transitionSpec,
 			animations,
