+++ packages/react-native-screen-transitions/src/shared/hooks/lifecycle/use-open-transition.ts
@@ -1,7 +1,7 @@
 import { useLayoutEffect } from "react";
 import type { BaseDescriptor } from "../../providers/screen/keys.provider";
 import type { AnimationStoreMap } from "../../stores/animation.store";
-import { startScreenTransition } from "../../utils/animation/start-screen-transition";
+import { animateToProgress } from "../../utils/animation/animate-to-progress";
 import { useHighRefreshRate } from "../animation/use-high-refresh-rate";
 
 /**
@@ -39,12 +39,11 @@ export function useOpenTransition(
 		const targetProgress = getInitialProgress(snapPoints, initialSnapIndex);
 
 		activateHighRefreshRate();
-		startScreenTransition({
-			target: "open",
+		animateToProgress({
+			target: targetProgress ?? "open",
 			spec: current.options.transitionSpec,
 			animations,
 			onAnimationFinish: deactivateHighRefreshRate,
-			targetProgress,
 		});
 	}, []);
